pyproject.toml
[project]
name = "job-market-analytics"
version = "0.1.0"
description = "Job market and GitHub trends analytics pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
    "snowflake-connector-python>=3.14.0",
    "dbt-snowflake>=1.10.0",
    "streamlit>=1.44.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]
